Avoiding code duplication in Vendor and Customer by adding parent class Partner
