+++ src/main/java/org/myshop/shop/model/Customer.java
@@ -1,20 +1,8 @@
 package org.myshop.shop.model;
 
-public class Customer {
+public class Customer extends Partner {
 
-    private String id = "";
-    private String name = "";
-    
-    public String getId() {
-        return id;
-    }
-    public void setId(String id) {
-        this.id = id;
-    }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    } 
+  public Customer() {
+	  super();
+  }
 }
