+++ src/main/java/org/myshop/shop/model/Partner.java
@@ -1,0 +1,23 @@
+package org.myshop.shop.model;
+
+public class Partner {
+	
+	private String id= " ";
+	private String name= " ";
+	
+	public Partner() {}
+	
+	public String getId() {
+		return id;
+	}
+	public void setId(String id) {
+		this.id = id;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+}
